+++ models/MessageEntry.php
@@ -3,8 +3,14 @@
 namespace humhub\modules\mail\models;
 
 use DateTime;
+use humhub\modules\content\widgets\richtext\AbstractRichText;
+use humhub\modules\content\widgets\richtext\converter\RichTextToEmailHtmlConverter;
+use humhub\modules\content\widgets\richtext\RichText;
+use humhub\modules\like\interfaces\LikeNotificationInterface;
+use humhub\modules\mail\helpers\Url;
 use humhub\modules\user\models\User;
 use Yii;
+use yii\base\InvalidConfigException;
 
 /**
  * This class represents a message text within a conversation.
@@ -12,7 +18,7 @@ use Yii;
  * @package humhub.modules.mail.models
  * @since 0.5
  */
-class MessageEntry extends AbstractMessageEntry
+class MessageEntry extends AbstractMessageEntry implements LikeNotificationInterface
 {
     /**
      * @inheritdoc
@@ -60,5 +66,28 @@ class MessageEntry extends AbstractMessageEntry
             ->andWhere(['!=', 'id', $this->id])
             ->andWhere(['>=', 'created_at', $today])
             ->exists();
+    }
+
+    /**
+     * @inerhitdoc
+     * @throws InvalidConfigException
+     */
+    public function getLikeNotificationPlainTextPreview(): string
+    {
+        return Yii::t('MailModule.base', 'Message') . ' ' . RichText::convert($this->content, AbstractRichText::FORMAT_SHORTTEXT);
+    }
+
+    /**
+     * @inerhitdoc
+     * @throws InvalidConfigException
+     */
+    public function getLikeNotificationHtmlPreview(): string
+    {
+        return RichTextToEmailHtmlConverter::process($this->content);
+    }
+
+    public function getLikeNotificationUrl(bool $scheme = false): string
+    {
+        return Url::toMessenger($this->message, $scheme);
     }
 }